[GLUTEN-6950][CORE] Move specific rules into backend modules
